requirements.txt
django
djangorestframework
Pillow
imagehash
opencv-python
numpy
scikit-learn
celery[redis]